// helpers/types.ts
export interface SymbolMap<T> {
  [symbol: string]: T;
}

export type eNetwork = eEthereumNetwork;

export enum eEthereumNetwork {
  coverage = "coverage",
  hardhat = "hardhat",
  localhost = "localhost",
  develop = "develop",
  rinkeby = "rinkeby",
  kovan = "kovan",
  main = "main",
  ropsten = "ropsten",
  goerli = "goerli",
}

export enum ConfettiPools {
  proto = "proto",
}

export enum eContractid {
  MintableERC20 = "MintableERC20",
  MintableERC721 = "MintableERC721",
  LendPoolAddressesProvider = "LendPoolAddressesProvider",
  LendPoolAddressesProviderRegistry = "LendPoolAddressesProviderRegistry",
  LendPoolParametersProvider = "LendPoolParametersProvider",
  LendPoolConfigurator = "LendPoolConfigurator",
  ValidationLogic = "ValidationLogic",
  ReserveLogic = "ReserveLogic",
  NftLogic = "NftLogic",
  GenericLogic = "GenericLogic",
  SupplyLogic = "SupplyLogic",
  BorrowLogic = "BorrowLogic",
  LiquidateLogic = "LiquidateLogic",
  ConfiguratorLogic = "ConfiguratorLogic",
  LendPool = "LendPool",
  LendPoolLoan = "LendPoolLoan",
  ReserveOracle = "ReserveOracle",
  ReserveOracleImpl = "ReserveOracleImpl",
  NFTOracle = "NFTOracle",
  NFTOracleImpl = "NFTOracleImpl",
  Proxy = "Proxy",
  MockChainlinkOracle = "MockChainlinkOracle",
  MockNFTOracle = "MockNFTOracle",
  MockReserveOracle = "MockReserveOracle",
  InterestRate = "InterestRate",
  ConfettiUpgradeableProxy = "ConfettiUpgradeableProxy",
  ConfettiProxyAdminTest = "ConfettiProxyAdminTest",
  ConfettiProxyAdminPool = "ConfettiProxyAdminPool", //LendPool Contracts, etc Oracle(Reserve, NFT)
  ConfettiProxyAdminFund = "ConfettiProxyAdminFund", //Treasury Fundings, etc Collector
  WalletBalanceProvider = "WalletBalanceProvider",
  CToken = "CToken",
  DebtToken = "DebtToken",
  CNFT = "CNFT",
  MockCNFT = "MockCNFT",
  ConfettiProtocolDataProvider = "ConfettiProtocolDataProvider",
  IERC20Detailed = "IERC20Detailed",
  IERC721Detailed = "IERC721Detailed",
  FeeProvider = "FeeProvider",
  WETHGateway = "WETHGateway",
  WETHGatewayImpl = "WETHGatewayImpl",
  WETH = "WETH",
  WETHMocked = "WETHMocked",
  WPUNKSGateway = "WPUNKSGateway",
  WPUNKS = "WPUNKS",
  WPUNKSMocked = "WPUNKSMocked",
  SelfdestructTransferMock = "SelfdestructTransferMock",
  LendPoolImpl = "LendPoolImpl",
  LendPoolConfiguratorImpl = "LendPoolConfiguratorImpl",
  LendPoolLoanImpl = "LendPoolLoanImpl",
  CNFTRegistry = "CNFTRegistry",
  CNFTRegistryImpl = "CNFTRegistryImpl",
  CryptoPunksMarket = "CryptoPunksMarket",
  WrappedPunk = "WrappedPunk",
  PunkGateway = "PunkGateway",
  PunkGatewayImpl = "PunkGatewayImpl",
  MockIncentivesController = "MockIncentivesController",
  UIPoolDataProvider = "UIPoolDataProvider",
  ConfettiCollector = "ConfettiCollector",
  ConfettiCollectorImpl = "ConfettiCollectorImpl",
  TimelockControllerFast = "TimelockControllerFast",
  TimelockControllerSlow = "TimelockControllerSlow",
}
// ...
export type tEthereumAddress = string;
export type tStringTokenBigUnits = string; // 1 ETH, or 10e6 USDC or 10e18 DAI
export type tBigNumberTokenBigUnits = BigNumber;
export type tStringTokenSmallUnits = string; // 1 wei, or 1 basic unit of USDC, or 1 basic unit of DAI
export type tBigNumberTokenSmallUnits = BigNumber;

export interface iAssetCommon<T> {
  [key: string]: T;
}
export interface iAssetBase<T> {
  //BUSD: T;
  WETH: T;
  DAI: T;
  // USDC: T;
  //USDT: T;
}

export type iAssetsWithoutETH<T> = Omit<iAssetBase<T>, "ETH">;

export type iAssetsWithoutUSD<T> = Omit<iAssetBase<T>, "USD">;

export type iConfettiPoolAssets<T> = Pick<
  iAssetsWithoutUSD<T>,
  "WETH" | "DAI"
  //| 'BUSD'
  // | "USDC"
  //| 'USDT'
>;

export type iMultiPoolsAssets<T> = iAssetCommon<T> | iConfettiPoolAssets<T>;

export type iConfettiPoolTokens<T> = Omit<iConfettiPoolAssets<T>, "ETH">;

export type iAssetAggregatorBase<T> = iAssetsWithoutETH<T>;

export enum TokenContractId {
  WETH = "WETH",
  DAI = "DAI",
  //BUSD = 'BUSD',
  // USDC = "USDC",
  //USDT = 'USDT',
}

export interface iNftCommon<T> {
  [key: string]: T;
}
export interface iNftBase<T> {
  WPUNKS: T;
  BAYC: T;
  DOODLE: T;
  // COOL: T;
  // MEEBITS: T;
  // MAYC: T;
  // WOW: T;
  // CLONEX: T;
  // AZUKI: T;
  // KONGZ: T;
}

export type iMultiPoolsNfts<T> = iNftCommon<T> | iConfettiPoolNfts<T>;

export type iConfettiPoolNfts<T> = iNftBase<T>;

export type iNftAggregatorBase<T> = iNftBase<T>;

export enum NftContractId {
  WPUNKS = "WPUNKS",
  BAYC = "BAYC",
  DOODLE = "DOODLE",
  // COOL = "COOL",
  // MEEBITS = "MEEBITS",
  // MAYC = "MAYC",
  // WOW = "WOW",
  // CLONEX = "CLONEX",
  // AZUKI = "AZUKI",
  // KONGZ = "KONGZ",
}

export interface IReserveParams
  extends IReserveBorrowParams,
    IReserveCollateralParams {
  cTokenImpl: eContractid;
  reserveFactor: string;
  strategy: IInterestRateStrategyParams;
}

export interface INftParams extends INftBuyItNowParams, INftCollateralParams {
  cNFTImpl: eContractid;
}

export interface IInterestRateStrategyParams {
  name: string;
  optimalUtilizationRate: string;
  baseVariableBorrowRate: string;
  variableRateSlope1: string;
  variableRateSlope2: string;
}

export interface IReserveBorrowParams {
  borrowingEnabled: boolean;
  reserveDecimals: string;
}

export interface IReserveCollateralParams {
  baseLTVAsCollateral: string;
  liquidationThreshold: string;
  liquidationBonus: string;
}

export interface INftCollateralParams {
  baseLTVAsCollateral: string;
  liquidationThreshold: string;
  liquidationBonus: string;
}

export interface INftBuyItNowParams {
  redeemDuration: string;
  buyItNowDuration: string;
  redeemFine: string;
  redeemThreshold: string;
  minBuyItNowFine: string;
}

export type iParamsPerNetwork<T> = iEthereumParamsPerNetwork<T>;

export interface iParamsPerNetworkAll<T> extends iEthereumParamsPerNetwork<T> {}

export interface iEthereumParamsPerNetwork<T> {
  [eEthereumNetwork.coverage]: T;
  [eEthereumNetwork.hardhat]: T;
  [eEthereumNetwork.localhost]: T;
  [eEthereumNetwork.develop]: T;
  [eEthereumNetwork.rinkeby]: T;
  [eEthereumNetwork.kovan]: T;
  [eEthereumNetwork.main]: T;
  [eEthereumNetwork.ropsten]: T;
  [eEthereumNetwork.goerli]: T;
}

export interface iParamsPerPool<T> {
  [ConfettiPools.proto]: T;
}

export interface iBasicDistributionParams {
  receivers: string[];
  percentages: string[];
}

export interface ObjectString {
  [key: string]: string;
}

export interface IProtocolGlobalConfig {
  MockUsdPrice: string;
  UsdAddress: tEthereumAddress;
  NilAddress: tEthereumAddress;
  OneAddress: tEthereumAddress;
}

export interface IMocksConfig {
  CNFTNamePrefix: string;
  CNFTSymbolPrefix: string;
  AllAssetsInitialPrices: iAssetBase<string>;
  AllNftsInitialPrices: iNftBase<string>;
}

export interface ICommonConfiguration {
  MarketId: string;
  CTokenNamePrefix: string;
  CTokenSymbolPrefix: string;
  DebtTokenNamePrefix: string;
  DebtTokenSymbolPrefix: string;

  ProviderId: number;
  ProtocolGlobalParams: IProtocolGlobalConfig;
  Mocks: IMocksConfig;

  ProxyAdminPool: iParamsPerNetwork<tEthereumAddress | undefined>;
  ProxyAdminFund: iParamsPerNetwork<tEthereumAddress | undefined>;

  CNFTRegistry: iParamsPerNetwork<tEthereumAddress | undefined>;

  ProviderRegistry: iParamsPerNetwork<tEthereumAddress | undefined>;
  ProviderRegistryOwner: iParamsPerNetwork<tEthereumAddress | undefined>;

  ReserveOracle: iParamsPerNetwork<tEthereumAddress | undefined>;
  NFTOracle: iParamsPerNetwork<tEthereumAddress | undefined>;

  PoolAdmin: iParamsPerNetwork<tEthereumAddress | undefined>;
  PoolAdminIndex: number;
  EmergencyAdmin: iParamsPerNetwork<tEthereumAddress | undefined>;
  EmergencyAdminIndex: number;

  ReserveAggregators: iParamsPerNetwork<ITokenAddress>;
  ReserveAssets: iParamsPerNetwork<SymbolMap<tEthereumAddress>>;
  ReservesConfig: iMultiPoolsAssets<IReserveParams>;
  NftsAssets: iParamsPerNetwork<SymbolMap<tEthereumAddress>>;
  NftsConfig: iMultiPoolsNfts<INftParams>;

  WrappedNativeToken: iParamsPerNetwork<tEthereumAddress>;

  CryptoPunksMarket: iParamsPerNetwork<tEthereumAddress>;
  WrappedPunkToken: iParamsPerNetwork<tEthereumAddress>;

  ReserveFactorTreasuryAddress: iParamsPerNetwork<tEthereumAddress>;
  IncentivesController: iParamsPerNetwork<tEthereumAddress>;
  DebtTokenImplementation?: iParamsPerNetwork<tEthereumAddress>;

  OracleQuoteCurrency: string;
  OracleQuoteUnit: string;
}

export interface IConfettiConfiguration extends ICommonConfiguration {
  ReservesConfig: iConfettiPoolAssets<IReserveParams>;
  NftsConfig: iConfettiPoolNfts<INftParams>;
}

export interface ITokenAddress {
  [token: string]: tEthereumAddress;
}